# app/docs/swagger.json
{
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "swagger": "2.0",
    "info": {
...
        "title": "pint-server",
        "contact": {},
        "license": {
            "name": "MIT"
        }
    },
    "paths": {
...
        "/admin/parties": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "Create a new party",
                "parameters": [
                    {
                        "description": "Party details",
                        "name": "party",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.PartyRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/handlers.PartyResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/admin/parties/": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "Get party by party code",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Code list provider",
                        "name": "code_list_provider",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Code list name (optional)",
                        "name": "code_list_name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Party code",
                        "name": "party_code",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.PartyResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Party not found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/admin/parties/{partyID}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "Get party by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Party ID",
                        "name": "partyID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.PartyResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid party ID",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Party not found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "Update an existing party",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Party ID",
                        "name": "partyID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Party details",
                        "name": "party",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.PartyRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.PartyResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Party not found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/admin/parties/{partyID}/codes": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "Add an identifying code to a party",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Party ID",
                        "name": "partyID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Code details",
                        "name": "code",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.PartyIdentifyingCodeRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/handlers.PartyIdentifyingCodeResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/admin/parties/{partyName}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "Get party by party name",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Party name",
                        "name": "partyName",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.PartyResponse"
                        }
                    },
                    "404": {
                        "description": "Party not found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handlers.PartyIdentifyingCodeRequest": {
            "type": "object",
            "properties": {
                "code_list_name": {
                    "type": "string"
                },
                "code_list_provider": {
                    "type": "string"
                },
                "party_code": {
                    "type": "string"
                }
            }
        },
        "handlers.PartyIdentifyingCodeResponse": {
            "type": "object",
            "properties": {
                "code_list_name": {
                    "type": "string"
                },
                "code_list_provider": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "party_code": {
                    "type": "string"
                },
                "party_id": {
                    "type": "string"
                }
            }
        },
        "handlers.PartyRequest": {
            "type": "object",
            "properties": {
                "active": {
                    "type": "boolean"
                },
                "party_name": {
                    "type": "string"
                }
            }
        },
        "handlers.PartyResponse": {
            "type": "object",
            "properties": {
                "active": {
                    "type": "boolean"
                },
                "id": {
                    "type": "string"
                },
                "party_name": {
                    "type": "string"
                }
            }
        },
...
    },
...
}